fork bedrock request specifically for now to avoid interface changes
