+++ crates/braintrust-llm-router/src/router.rs
@@ -1,6 +1,4 @@
 use std::collections::HashMap;
-use std::future::Future;
-use std::pin::Pin;
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::time::sleep;
@@ -13,13 +11,12 @@ use bytes::Bytes;
 use crate::auth::AuthConfig;
 use crate::catalog::{load_catalog_from_disk, ModelCatalog, ModelResolver, ModelSpec};
 use crate::error::{Error, Result};
-use crate::providers::{ClientHeaders, Provider};
+use crate::providers::{
+    prepare_bedrock_request, requires_bedrock_request_preparation, ClientHeaders, Provider,
+};
 use crate::retry::{RetryPolicy, RetryStrategy};
 use crate::streaming::{transform_stream, ResponseStream};
-use lingua::processing::{adapter_for_format, adapters};
 use lingua::serde_json::Value;
-use lingua::universal::message::{Message, UserContent, UserContentPart};
-use lingua::util::media::MediaBlock;
 use lingua::ProviderFormat;
 use lingua::{TransformError, TransformResult};
 
@@ -112,134 +109,13 @@ type ResolvedRoute<'a> = (
     RetryStrategy,
 );
 
-fn is_remote_image_url(value: &str) -> bool {
-    value.starts_with("http://") || value.starts_with("https://")
-}
-
-const BEDROCK_REMOTE_MEDIA_MAX_BYTES: usize = 5 * 1024 * 1024;
-
-async fn fetch_remote_image_as_base64(url: &str) -> Result<MediaBlock> {
-    lingua::util::media::convert_media_to_base64(url, None, Some(BEDROCK_REMOTE_MEDIA_MAX_BYTES))
-        .await
-        .map_err(|e| Error::InvalidRequest(format!("failed to fetch image URL {url}: {e}")))
-}
-
-type FetchMediaFuture<'a> = Pin<Box<dyn Future<Output = Result<MediaBlock>> + Send + 'a>>;
-
-// Preserve the legacy proxy behavior that fetched remote image URLs before
-// Bedrock request translation, since Converse and Bedrock Anthropic expect
-// inline media rather than raw remote URLs.
-fn should_inline_remote_image_urls(format: ProviderFormat) -> bool {
-    matches!(
-        format,
-        ProviderFormat::BedrockAnthropic | ProviderFormat::Converse
-    )
-}
-
-async fn inline_remote_image_urls_with_fetch<F>(
-    request: &mut lingua::UniversalRequest,
-    mut fetch: F,
-) -> Result<()>
-where
-    F: for<'a> FnMut(&'a str) -> FetchMediaFuture<'a>,
-{
-    for message in &mut request.messages {
-        let content = match message {
-            Message::System { content }
-            | Message::Developer { content }
-            | Message::User { content } => content,
-            Message::Assistant { .. } | Message::Tool { .. } => continue,
-        };
-
-        let UserContent::Array(parts) = content else {
-            continue;
-        };
-
-        for part in parts {
-            let UserContentPart::Image {
-                image, media_type, ..
-            } = part
-            else {
-                continue;
-            };
-
-            let Some(url) = image.as_str() else {
-                continue;
-            };
-
-            if !is_remote_image_url(url) {
-                continue;
-            }
-
-            let media_block = fetch(url).await?;
-            *image = lingua::serde_json::Value::String(media_block.data);
-            *media_type = Some(media_block.media_type);
-        }
-    }
-
-    Ok(())
-}
-
-async fn prepare_request_with_remote_image_inlining<F>(
-    body: Bytes,
-    spec: &ModelSpec,
-    format: ProviderFormat,
-    fetch: F,
-) -> Result<Bytes>
-where
-    F: for<'a> FnMut(&'a str) -> FetchMediaFuture<'a>,
-{
-    let payload: lingua::serde_json::Value = match lingua::serde_json::from_slice(&body) {
-        Ok(payload) => payload,
-        Err(err) => {
-            return Err(TransformError::DeserializationFailed(err.to_string()).into());
-        }
-    };
-
-    let source_adapter = match adapters()
-        .iter()
-        .map(|adapter| adapter.as_ref())
-        .find(|adapter| adapter.detect_request(&payload))
-    {
-        Some(adapter) => adapter,
-        None => return Err(TransformError::UnableToDetectFormat.into()),
-    };
-
-    if source_adapter.format() == format {
-        return Ok(body);
-    }
-
-    let mut request = match source_adapter.request_to_universal(payload) {
-        Ok(request) => request,
-        Err(err) => return Err(err.into()),
-    };
-
-    inline_remote_image_urls_with_fetch(&mut request, fetch).await?;
-
-    if request.model.is_none() {
-        request.model = Some(spec.model.clone());
-    }
-
-    let target_adapter =
-        adapter_for_format(format).ok_or(TransformError::UnsupportedTargetFormat(format))?;
-    target_adapter.apply_defaults(&mut request);
-    let prepared = target_adapter.request_from_universal(&request)?;
-
-    lingua::serde_json::to_vec(&prepared)
-        .map(Bytes::from)
-        .map_err(Error::LinguaJson)
-}
-
 async fn prepare_provider_request(
     body: Bytes,
     spec: &ModelSpec,
     format: ProviderFormat,
 ) -> Result<Bytes> {
-    if should_inline_remote_image_urls(format) {
-        return prepare_request_with_remote_image_inlining(body, spec, format, |url| {
-            Box::pin(fetch_remote_image_as_base64(url))
-        })
-        .await;
+    if requires_bedrock_request_preparation(format) {
+        return prepare_bedrock_request(body, spec, format).await;
     }
 
     match lingua::transform_request(body.clone(), format, Some(&spec.model)) {
@@ -746,7 +622,6 @@ mod tests {
     use crate::streaming::RawResponseStream;
     use async_trait::async_trait;
     use reqwest::header::HeaderMap;
-    use std::sync::atomic::{AtomicUsize, Ordering};
 
     struct FakeProvider {
         name: &'static str,
@@ -840,229 +715,12 @@ mod tests {
         spec
     }
 
-    fn bedrock_spec(model: &str, format: ProviderFormat) -> ModelSpec {
-        ModelSpec {
-            model: model.to_string(),
-            format,
-            flavor: ModelFlavor::Chat,
-            display_name: None,
-            parent: None,
-            input_cost_per_mil_tokens: None,
-            output_cost_per_mil_tokens: None,
-            input_cache_read_cost_per_mil_tokens: None,
-            multimodal: None,
-            reasoning: None,
-            max_input_tokens: None,
-            max_output_tokens: None,
-            supports_streaming: true,
-            extra: Default::default(),
-            available_providers: Default::default(),
-        }
-    }
-
     fn dummy_auth() -> AuthConfig {
         AuthConfig::ApiKey {
             key: "test".into(),
             header: Some("authorization".into()),
             prefix: Some("Bearer".into()),
         }
-    }
-
-    #[test]
-    fn should_inline_remote_image_urls_matches_legacy_proxy_formats() {
-        assert!(should_inline_remote_image_urls(
-            ProviderFormat::BedrockAnthropic
-        ));
-        assert!(should_inline_remote_image_urls(ProviderFormat::Converse));
-        assert!(!should_inline_remote_image_urls(ProviderFormat::Anthropic));
-        assert!(!should_inline_remote_image_urls(
-            ProviderFormat::ChatCompletions
-        ));
-        assert!(!should_inline_remote_image_urls(ProviderFormat::Responses));
-        assert!(!should_inline_remote_image_urls(ProviderFormat::Google));
-    }
-
-    #[tokio::test]
-    async fn prepare_request_passes_through_same_format_converse_without_fetch() {
-        let body = Bytes::from(
-            lingua::serde_json::to_vec(&lingua::serde_json::json!({
-                "modelId": "anthropic.claude-3-haiku-20240307-v1:0",
-                "messages": [{
-                    "role": "user",
-                    "content": [{"text": "Hello"}]
-                }]
-            }))
-            .unwrap(),
-        );
-
-        let prepared = prepare_request_with_remote_image_inlining(
-            body.clone(),
-            &bedrock_spec(
-                "anthropic.claude-3-haiku-20240307-v1:0",
-                ProviderFormat::Converse,
-            ),
-            ProviderFormat::Converse,
-            |_url| {
-                Box::pin(async {
-                    panic!("fetch should not be called for same-format converse requests");
-                })
-            },
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(prepared, body);
-    }
-
-    #[tokio::test]
-    async fn prepare_request_inlines_remote_chat_image_for_converse() {
-        let body = Bytes::from(
-            lingua::serde_json::to_vec(&lingua::serde_json::json!({
-                "model": "claude-sonnet-4-5-20250929",
-                "messages": [{
-                    "role": "user",
-                    "content": [
-                        {"type": "text", "text": "What is this?"},
-                        {"type": "image_url", "image_url": {"url": "https://example.com/image.jpg"}}
-                    ]
-                }]
-            }))
-            .unwrap(),
-        );
-
-        let prepared = prepare_request_with_remote_image_inlining(
-            body,
-            &bedrock_spec(
-                "anthropic.claude-3-haiku-20240307-v1:0",
-                ProviderFormat::Converse,
-            ),
-            ProviderFormat::Converse,
-            |_url| {
-                Box::pin(async {
-                    Ok(MediaBlock {
-                        media_type: "image/jpeg".to_string(),
-                        data: "abcd".to_string(),
-                    })
-                })
-            },
-        )
-        .await
-        .unwrap();
-        let value: lingua::serde_json::Value = lingua::serde_json::from_slice(&prepared).unwrap();
-
-        let bytes = value
-            .pointer("/messages/0/content/1/image/source/bytes")
-            .and_then(|v| v.as_str())
-            .unwrap();
-        assert!(!bytes.is_empty());
-        assert_eq!(
-            value
-                .pointer("/messages/0/content/1/image/format")
-                .and_then(|v| v.as_str()),
-            Some("jpeg")
-        );
-    }
-
-    #[tokio::test]
-    async fn prepare_request_inlines_remote_responses_image_for_bedrock_anthropic() {
-        let body = Bytes::from(
-            lingua::serde_json::to_vec(&lingua::serde_json::json!({
-                "model": "claude-sonnet-4-5-20250929",
-                "input": [{
-                    "role": "user",
-                    "content": [
-                        {"type": "input_text", "text": "What is this?"},
-                        {
-                            "type": "input_image",
-                            "image_url": "https://example.com/image.jpg",
-                            "detail": "auto"
-                        }
-                    ]
-                }]
-            }))
-            .unwrap(),
-        );
-
-        let prepared = prepare_request_with_remote_image_inlining(
-            body,
-            &bedrock_spec(
-                "anthropic.claude-3-haiku-20240307-v1:0",
-                ProviderFormat::BedrockAnthropic,
-            ),
-            ProviderFormat::BedrockAnthropic,
-            |_url| {
-                Box::pin(async {
-                    Ok(MediaBlock {
-                        media_type: "image/jpeg".to_string(),
-                        data: "abcd".to_string(),
-                    })
-                })
-            },
-        )
-        .await
-        .unwrap();
-        let value: lingua::serde_json::Value = lingua::serde_json::from_slice(&prepared).unwrap();
-
-        assert_eq!(
-            value.get("anthropic_version").and_then(|v| v.as_str()),
-            Some("bedrock-2023-05-31")
-        );
-        assert_eq!(
-            value
-                .pointer("/messages/0/content/1/source/type")
-                .and_then(|v| v.as_str()),
-            Some("base64")
-        );
-        assert!(value
-            .pointer("/messages/0/content/1/source/data")
-            .and_then(|v| v.as_str())
-            .is_some_and(|v| !v.is_empty()));
-        assert_eq!(value.pointer("/messages/0/content/1/source/url"), None);
-    }
-
-    #[tokio::test]
-    async fn prepare_request_returns_invalid_request_when_remote_image_fetch_fails() {
-        let fetch_calls = Arc::new(AtomicUsize::new(0));
-        let body = Bytes::from(
-            lingua::serde_json::to_vec(&lingua::serde_json::json!({
-                "model": "claude-sonnet-4-5-20250929",
-                "messages": [{
-                    "role": "user",
-                    "content": [
-                        {"type": "text", "text": "What is this?"},
-                        {"type": "image_url", "image_url": {"url": "https://example.com/image.jpg"}}
-                    ]
-                }]
-            }))
-            .unwrap(),
-        );
-
-        let err = prepare_request_with_remote_image_inlining(
-            body,
-            &bedrock_spec(
-                "anthropic.claude-3-haiku-20240307-v1:0",
-                ProviderFormat::Converse,
-            ),
-            ProviderFormat::Converse,
-            {
-                let fetch_calls = Arc::clone(&fetch_calls);
-                move |url| {
-                    fetch_calls.fetch_add(1, Ordering::SeqCst);
-                    Box::pin(async move {
-                        Err(Error::InvalidRequest(format!(
-                            "failed to fetch image URL {url}: network error"
-                        )))
-                    })
-                }
-            },
-        )
-        .await
-        .expect_err("fetch failure should surface as InvalidRequest");
-
-        assert_eq!(fetch_calls.load(Ordering::SeqCst), 1);
-        assert!(
-            matches!(err, Error::InvalidRequest(ref msg) if msg.contains("failed to fetch image URL"))
-        );
     }
 
     #[test]
